Update app.js -- Use of Javascript object
I believe it is better to use a Javascript object (dictionary) to store the nicknames, as we can access them directly by their values (i.e. nicknames themselves).

Notations are much simpler and we also save us from using indexOf().


NOTE: This require a change in how we parse nicknames in index.html.